+++ project_6/include/findSalary.hpp
@@ -2,12 +2,13 @@
 #define FIND_SALARY_HPP
 
 #include "employee.hpp"
+#include <cstddef>
 #include <cstring>
 #include <iostream>
 #include <vector>
 #include "storageEmployees.hpp"
 
-std::vector<int> findSalary(const SafeVector& employees,
+std::vector<size_t> findSalary(const SafeVector& employees,
                             double salary);
 
 #endif
