+++ project_6/src/findDept.cpp
@@ -1,8 +1,9 @@
 #include "findDept.hpp"
+#include <cstddef>
 
-std::vector<int> findDept(const SafeVector& employees, int dept)
+std::vector<size_t> findDept(const SafeVector& employees, int dept)
 {
-    std::vector<int> idxs;
+    std::vector<size_t> idxs;
     for (size_t idx = 0; idx < employees.getSize(); idx++)
     {
         if (employees.at(idx).getDept() == dept)
