+++ project_6/src/findSalary.cpp
@@ -1,8 +1,9 @@
 #include "findSalary.hpp"
+#include <cstddef>
 
-std::vector<int> findSalary(const SafeVector& employees, double salary)
+std::vector<size_t> findSalary(const SafeVector& employees, double salary)
 {
-    std::vector<int> idxs;
+    std::vector<size_t> idxs;
     for (size_t idx = 0; idx < employees.getSize(); idx++)
     {
         if (employees.at(idx).getSalary() == salary)
